MLDB-1433 basic of new random forest with classifier function
