+++ ml/jml/committee.cc
@@ -171,12 +171,23 @@ optimized_predict_impl(int label,
 
     for (unsigned i = 0;  i < classifiers.size();  ++i) {
         if (weights[i] == 0.0) continue;
-        result = result
+        float vote = classifiers[i]->optimized_predict_impl(label, features, info, context);
+        cerr << "Vote: " << vote << endl;
+        //if (!isnanf(vote))
+        {
+            result = result
             + weights[i]
-            * classifiers[i]->optimized_predict_impl(label, features, info,
-                                                     context);
+            * vote;
+        }
+        //else
+       // {
+       //     result = result
+       //     + weights[i]
+      //      * 0.5f;
+       // }        
     }
 
+    cerr << "------------" << endl;
     return result;
 }
