+++ ml/jml/decision_tree.cc
@@ -153,6 +153,8 @@ struct LabelResults {
     JML_ALWAYS_INLINE
     void operator () (const Label_Dist & dist, float weight)
     {
+        ExcAssert(!isnanf(weight));
+        ExcAssert(!isnanf(dist[label]));
         result += weight * dist[label];
     }
 
@@ -281,6 +283,7 @@ predict_recursive_impl(const GetFeatures & get_features,
                        const Tree::Ptr & ptr,
                        double weight) const
 {
+    ExcAssert(!isnanf(weight));
     if (!ptr) return;
 
     if (!ptr.node()) {
@@ -300,8 +303,11 @@ predict_recursive_impl(const GetFeatures & get_features,
         predict_recursive_impl(get_features, results, node.child_false,
                                weights[false]);
     if (weights[MISSING] > 0.0)
+    {
+        ExcAssert(false);
         predict_recursive_impl(get_features, results, node.child_missing,
                                weights[MISSING]);
+    }
 }
 
 std::string
