+++ server/column_scope.cc
@@ -1,0 +1,2 @@
+
+#include "column_scope.h"
